+++ conversion/ft_puthex.c
@@ -1,9 +1,9 @@
 #include "../printf.h"
 
-int	ft_puthex(unsigned long long n, char *base)
+int	ft_puthex(unsigned int n, char *base)
 {
-	int	i;
-	int	len;
+	unsigned int	i;
+	int				len;
 
 	i = ft_strlen(base);
 	len = 0;
