+++ printf.h
@@ -20,8 +20,8 @@ int				ft_strlen(const char *str);
 int				ft_putstr(char *str);
 int				ft_putnbr(int nb);
 unsigned int	ft_putnbr_unsigned(unsigned int nb);
-int				ft_puthex(unsigned long long n, char *base);
-int				ft_putaddr(unsigned long long int n);
+int				ft_puthex(unsigned int n, char *base);
+int				ft_putaddr(unsigned int n);
 
 const char		*ft_search_arg(va_list arg, const char *format, t_lw *sc);
 const char		*ft_read_text(t_lw *sc, const char *format);
